requirements.txt
Flask
openai
ollama
flask 
flask-bcrypt